Dartboard Oomph setup
This adds an Oomph setup file so that it is easier for contributors to
contribute to the Dartboard project.
This is the very fist step, it basically only:

- Installs the Eclipse SDK
- Sets the Java heap space to 4GB.
- Installs the Tycho extra plug-in to avoid showing errors on the pom
file
- Activates the target platform
- Imports the Git repository
- Creates a working set for the projects included in the Dartboard git
repo
- And points to the setup file on the server

Due to Bug https://bugs.eclipse.org/bugs/show_bug.cgi?id=552315 I cannot
easily test this before pushing it.